Retrieve relevant data from LOC API
extracted only 3 values from the library of Congress API's JSON
response because the text recognition for the newspaper pages is very
poor and displaying the newspaper text is not helpful to the user in
its current garbled state. so they will be taken
to the website to choose how to search and view the returned search
results from the API. Limited search results to 5 using a for loop
so that users won't be combing through a lot of irrelevant data.