Display date and title of newspaper
Needed to fix the logic to render only part of the value of the `title`
key in the first 5 items of the `items` array returned by the
Library of Congress API so that it was clear to readers that it was a
newspaper title. Also arranged the string of numbers stored in
the `date` key of that same `items` array for the 5 items so that
users could easily tell it was a date.