Return human-readable text in place of API markup
I had to add many .replace() methods to eliminate all the markup in the
Etymology information strings returned by the Merriam Webster API so
the user wouldn't be confused and distracted by it. I either removed
the markup completely for simplicity's sake, or added simple html to
reflect the formatting the markup was suggesting.